requirements.txt
Flask==2.1.1
Flask-JWT-Extended==4.3.1
pandas==1.3.5
requests==2.27.1
SQLAlchemy==1.4.31
plotly==5.6.0
python-dotenv==0.20.0
Werkzeug==2.0.3  # For password hashing and security
